pyproject.toml
[project]
name = "llm-mas-ecrp"
version = "0.1.0"
description = ""
authors = [
    {name = "pengkangzhen",email = "pengkangzhen@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "huggingface-hub (>=0.20.0,<1.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "openai (>=2.7.2,<3.0.0)",
    "gurobipy (==11.0.3)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "llm_mas_ecrp", from = "src"}]

[tool.poetry.scripts]
load-nlp4lp = "scripts.load_nlp4lp_csv:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
